requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
marshmallow==3.20.1
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
